requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.27
pydantic>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pytest>=7.4.0
pytest-cov>=4.1.0
pylint>=3.0.3
python-dotenv>=1.0.0
requests>=2.31.0 